refactor(app): transformations w/ hardware hooks
We have this top-level directory in app called resources/, which is for
gathering hooks whose main job is to interact with the robot, but whose
actions are too specific to the app for being useful in the general
react-api-client.

We also have a bunch of hooks and free functions whose job is to take
data from those hooks, and transform them in some way. For instance, all
the things that get interesting information from lists of commands.

Today, those things are all inside some organism or page, and that's
sometimes fine.... but some of them are used across other organisms or
pages, and that's gross. So let's put them in a new top-level directory transformations/
