+++ app/src/organisms/Devices/ProtocolRun/SetupLabware/__tests__/getNestedLabwareInfo.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'vitest'
 import { mockDefinition } from '../../../../../redux/custom-labware/__fixtures__'
 import { getNestedLabwareInfo } from '../getNestedLabwareInfo'
 import type { RunTimeCommand } from '@opentrons/shared-data'
-import type { LabwareSetupItem } from '../../../../../pages/Desktop/Protocols/utils'
+import type { LabwareSetupItem } from '../../../../../transformations/commands'
 
 const MOCK_LABWARE_ID = 'mockLabwareId'
 const MOCK_OTHER_LABWARE_ID = 'mockOtherLabwareId'
