+++ app/src/organisms/ProtocolDetails/index.tsx
@@ -67,7 +67,7 @@ import {
   getAnalysisStatus,
   getProtocolDisplayName,
 } from '../ProtocolsLanding/utils'
-import { getProtocolUsesGripper } from '../ODD/ProtocolSetup/ProtocolSetupInstruments/utils'
+import { getProtocolUsesGripper } from '../../transformations/commands'
 import { ProtocolOverflowMenu } from '../ProtocolsLanding/ProtocolOverflowMenu'
 import { ProtocolStats } from './ProtocolStats'
 import { ProtocolLabwareDetails } from './ProtocolLabwareDetails'
