+++ app/src/pages/ODD/ProtocolDetails/Deck.tsx
@@ -10,7 +10,7 @@ import { getLabwareDefURI } from '@opentrons/shared-data'
 
 import { LabwareStackModal } from '../../../organisms/Devices/ProtocolRun/SetupLabware/LabwareStackModal'
 import { SingleLabwareModal } from '../../../organisms/ODD/ProtocolSetup/ProtocolSetupLabware/SingleLabwareModal'
-import { getLabwareSetupItemGroups } from '../../../pages/Desktop/Protocols/utils'
+import { getLabwareSetupItemGroups } from '../../../transformations/commands'
 
 import type {
   LabwareDefinition2,
