+++ app/src/pages/ODD/QuickTransferDetails/__tests__/Hardware.test.tsx
@@ -12,6 +12,7 @@ import { i18n } from '../../../../i18n'
 import { useRequiredProtocolHardware } from '../../../../pages/Desktop/Protocols/hooks'
 import { Hardware } from '../Hardware'
 
+vi.mock('../../../../transformations/commands')
 vi.mock('../../../../pages/Desktop/Protocols/hooks')
 vi.mock('../../../../redux/config')
