+++ app/src/pages/ODD/RobotDashboard/__tests__/RobotDashboard.test.tsx
@@ -12,7 +12,7 @@ import {
   EmptyRecentRun,
 } from '../../../../organisms/ODD/RobotDashboard'
 import { Navigation } from '../../../../organisms/Navigation'
-import { useMissingProtocolHardware } from '../../../../pages/Desktop/Protocols/hooks'
+import { useMissingProtocolHardware } from '../../../../transformations/commands'
 import { getOnDeviceDisplaySettings } from '../../../../redux/config'
 import { WelcomeModal } from '../WelcomeModal'
 import { RobotDashboard } from '..'
@@ -34,7 +34,7 @@ vi.mock('@opentrons/react-api-client')
 vi.mock('../../../../organisms/ODD/RobotDashboard/EmptyRecentRun')
 vi.mock('../../../../organisms/ODD/RobotDashboard/RecentRunProtocolCarousel')
 vi.mock('../../../../organisms/Navigation')
-vi.mock('../../../../pages/Desktop/Protocols/hooks')
+vi.mock('../../../../transformations/commands')
 vi.mock('../../../../redux/config')
 vi.mock('../WelcomeModal')
 vi.mock('../../../../resources/runs')
